requirements.txt
nest_asyncio
argparse
requests_html
bs4
pytube3
